argflow/requirements.txt
appdirs==1.4.3
coverage==5.3
Keras==2.4.3
matplotlib==3.2.2
networkx==2.4
numpy==1.19.2
pandas==1.0.5
Pillow==7.0.0
requests==2.24.0
sklearn==0.0
tables==3.6.1
tensorflow==2.3.1
tf-keras-vis==0.5.3
scipy==1.3